added support for incrementation != 1, ranges belows zero

+++ include/ppbar/common.hpp
@@ -6,7 +6,7 @@
 namespace ppbar
 {
 
-typedef size_t T;
+typedef int T;
